refactor: renaming infrastructure modules
