+++ src/domain/checkout/repository/order-repository.interface.ts
@@ -1,4 +1,4 @@
 import Order from "../entity/order";
 import RepositoryInterface from "../../@shared/repository/repository-interface";
 
-export default interface ProductRepositoryInterface extends RepositoryInterface<Order> {}
+export default interface OderRepositoryInterface extends RepositoryInterface<Order> {}
