+++ src/infrastructure/customer/repository/sequelize/customer.repository.spec.ts
@@ -1,7 +1,7 @@
 import { Sequelize } from "sequelize-typescript";
-import Customer from "../../domain/customer/entity/customer"; 
-import Address from "../../domain/customer/value-object/address";
-import CustomerModel from "../db/sequelize/model/customer.model";
+import Customer from "../../../../domain/customer/entity/customer"; 
+import Address from "../../../../domain/customer/value-object/address";
+import CustomerModel from "./customer.model";
 import CustomerRepository from "./customer.repository";
 
 describe("Customer repository test", () => {
