+++ src/infrastructure/customer/repository/sequelize/customer.repository.ts
@@ -1,7 +1,7 @@
-import Customer from "../../domain/customer/entity/customer";
-import Address from "../../domain/customer/value-object/address";
-import CustomerRepositoryInterface from "../../domain/customer/repository/customer-reposity.interface";
-import CustomerModel from "../db/sequelize/model/customer.model";
+import Customer from "../../../../domain/customer/entity/customer";
+import Address from "../../../../domain/customer/value-object/address";
+import CustomerRepositoryInterface from "../../../../domain/customer/repository/customer-reposity.interface";
+import CustomerModel from "./customer.model";
 
 export default class CustomerRepository implements CustomerRepositoryInterface {
   async create(entity: Customer): Promise<void> {
