+++ src/infrastructure/order/repository/sequelize/order-item.model.ts
@@ -1,5 +1,5 @@
 import { Table, Model, PrimaryKey, Column, ForeignKey, BelongsTo } from "sequelize-typescript";
-import ProductModel from "./product.model";
+import ProductModel from "../../../product/repository/sequelize/product.model";
 import OrderModel from "./order.model";
 
 @Table({
