+++ src/infrastructure/order/repository/sequelize/order.model.ts
@@ -1,11 +1,12 @@
 import { Table, Model, PrimaryKey, Column, ForeignKey, BelongsTo, HasMany } from "sequelize-typescript";
-import CustomerModel from "./customer.model";
+import CustomerModel from "../../../customer/repository/sequelize/customer.model";
 import OrderItemModel from "./order-item.model";
 
 @Table({
   tableName: "orders",
   timestamps: false,
 })
+
 export default class OrderModel extends Model {
   @PrimaryKey
   @Column
